requirements.txt
PyQt6>=6.6.0
python-chess>=1.999
aiohttp>=3.9.0
qasync>=0.27.0
